requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements.in
#
base58==2.1.1
    # via -r ./requirements.in
blinker==1.6.2
    # via flask
cachetools==5.3.0
    # via google-auth
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
diskcache==5.6.1
    # via -r ./requirements.in
flask==2.3.2
    # via -r ./requirements.in
google-api-core[grpc]==2.11.0
    # via google-cloud-texttospeech
google-auth==2.18.1
    # via google-api-core
google-cloud-texttospeech==2.14.1
    # via -r ./requirements.in
googleapis-common-protos==1.59.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.54.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.54.2
    # via google-api-core
idna==3.4
    # via requests
imgkit==1.2.3
    # via -r ./requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
proto-plus==1.22.2
    # via google-cloud-texttospeech
protobuf==4.23.1
    # via
    #   google-api-core
    #   google-cloud-texttospeech
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
requests==2.31.0
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   imgkit
typing-extensions==4.6.1
    # via -r ./requirements.in
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
werkzeug==2.3.4
    # via flask
yatl==20230507.3
    # via -r ./requirements.in
